+++ config/web.php
@@ -4,8 +4,8 @@ use app\models\user\UserMaster;
 use sbs\behaviors\LastVisitBehavior;
 
 return yii\helpers\ArrayHelper::merge(
-    require __DIR__ . '/main.php',
-    require __DIR__ . '/main-local.php',
+    require __DIR__ . '/common.php',
+    require __DIR__ . '/common-local.php',
     [
         'id' => 'basic',
         'components' => [