+++ models/user/PasswordResetRequestForm.php
@@ -4,6 +4,7 @@ namespace app\models\user;
 
 use Yii;
 use yii\base\Model;
+use yii\base\Exception;
 
 /**
  * Class PasswordResetRequestForm
@@ -26,8 +27,7 @@ class PasswordResetRequestForm extends Model
                 'email',
                 'exist',
                 'targetClass' => UserMaster::class,
-                'filter' => ['status' => UserMaster::STATUS_ACTIVE],
-                'message' => 'There is no user with this email address.'
+                'message' => Yii::t('app', 'There is no user with this email address.')
             ],
         ];
     }
@@ -36,34 +36,41 @@ class PasswordResetRequestForm extends Model
      * Sends an email with a link, for resetting the password.
      *
      * @return bool whether the email was send
-     *
-     * @throws \yii\base\Exception
      */
     public function sendEmail()
     {
-        /* @var $user UserMaster */
-        $user = UserMaster::findOne(['status' => UserMaster::STATUS_ACTIVE, 'email' => $this->email]);
-
-        if (!$user instanceof UserMaster) {
+        if (!$this->validate()) {
             return false;
         }
 
-        if (!UserMaster::isPasswordResetTokenValid($user->password_reset_token)) {
-            $user->generatePasswordResetToken();
-            if (!$user->save()) {
-                return false;
+        try {
+            /* @var $user UserMaster */
+            $user = UserMaster::findOne(['email' => $this->email]);
+            if ($user->status !== UserMaster::STATUS_ACTIVE) {
+                throw new Exception(Yii::t('app',
+                    'The user is not active. We cannot send email to this type of user.'));
             }
-        }
 
-        return Yii::$app
-            ->mailer
-            ->compose(
-                ['html' => 'password-reset-request-html', 'text' => 'password-reset-request-text'],
-                ['user' => $user]
-            )
-            ->setFrom([Yii::$app->params['supportEmail'] => Yii::$app->name . ' Robot'])
-            ->setTo($this->email)
-            ->setSubject(Yii::t('app', 'Password reset on {portal}', ['portal' => Yii::$app->name]))
-            ->send();
+            if (!UserMaster::isPasswordResetTokenValid($user->password_reset_token)) {
+                $user->generatePasswordResetToken();
+                $user->save(false);
+            }
+
+            return Yii::$app
+                ->mailer
+                ->compose(
+                    ['html' => 'password-reset-request-html', 'text' => 'password-reset-request-text'],
+                    ['user' => $user]
+                )
+                ->setFrom([Yii::$app->params['supportEmail'] => Yii::$app->name . ' Robot'])
+                ->setTo($this->email)
+                ->setSubject(Yii::t('app', 'Password reset on {portal}', ['portal' => Yii::$app->name]))
+                ->send();
+
+        } catch (Exception $exception) {
+            $this->addError('email', $exception->getMessage());
+
+            return false;
+        }
     }
 }