+++ tests/_support/FunctionalTester.php
@@ -15,9 +15,18 @@
  * @method \Codeception\Lib\Friend haveFriend($name, $actorClass = NULL)
  *
  * @SuppressWarnings(PHPMD)
-*/
+ */
 class FunctionalTester extends \Codeception\Actor
 {
     use _generated\FunctionalTesterActions;
 
+    public function seeValidationError($message)
+    {
+        $this->see($message, '.help-block');
+    }
+
+    public function dontSeeValidationError($message)
+    {
+        $this->dontSee($message, '.help-block');
+    }
 }